feat(auth): implement role-based authorization for role management actions; update permissions in seeder

+++ app/Http/Controllers/RoleController.php
@@ -3,22 +3,21 @@
 namespace App\Http\Controllers;
 
 use App\Models\User;
-use Illuminate\Http\Request;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Hash;
 use App\Services\Interfaces\RoleInterface;
 use App\Http\Requests\Role\StoreRoleRequest;
 use App\Http\Requests\Role\UpdateRoleRequest;
 
 class RoleController extends Controller
 {
+    use AuthorizesRequests;
+
     public function __construct(private RoleInterface $roleRepository) {}
 
     public function index()
     {
-        if (!Auth::user()->hasRole('owner')) {
-            return abort(403, 'Unauthorized action.');
-        }
+        $this->authorize('role.index');
 
         $users = $this->roleRepository->getAllRoles();
 
@@ -27,9 +26,7 @@ class RoleController extends Controller
 
     public function create()
     {
-        if (!Auth::user()->hasRole('owner')) {
-            return abort(403, 'Unauthorized action.');
-        }
+        $this->authorize('role.create');
 
         return view('pages.role.create');
     }
@@ -45,9 +42,7 @@ class RoleController extends Controller
 
     public function edit(User $role)
     {
-        if (!Auth::user()->hasRole('owner')) {
-            return abort(403, 'Unauthorized action.');
-        }
+        $this->authorize('role.edit');
 
         return view('pages.role.edit', compact('role'));
     }
@@ -63,6 +58,8 @@ class RoleController extends Controller
 
     public function destroy(User $role)
     {
+        $this->authorize('role.destroy');
+
         $role->delete();
 
         return redirect()->route('roles.index')->withSuccess('Berhasil dihapus');
